+++ samples/Extensions/Timer/TimerFunction.cs
@@ -9,12 +9,12 @@ namespace SampleApp
     public static class TimerFunction
     {
         //<docsnippet_fixed_delay_retry_example>
-        [Function("TimerFunction")]
+        [Function(nameof(TimerFunction))]
         [FixedDelayRetry(5, "00:00:10")]
         public static void Run([TimerTrigger("0 */5 * * * *")] TimerInfo timerInfo,
             FunctionContext context)
         {
-            var logger = context.GetLogger("TimerFunction");
+            var logger = context.GetLogger(nameof(TimerFunction));
             logger.LogInformation($"Function Ran. Next timer schedule = {timerInfo.ScheduleStatus.Next}");
         }
         //</docsnippet_fixed_delay_retry_example>