+++ src/DotNetWorker.Core/Converters/Converter/IConverterContextFactory.cs
@@ -17,16 +17,7 @@ namespace Microsoft.Azure.Functions.Worker.Converters
         /// <param name="targetType">The target type.</param>
         /// <param name="source">The source.</param>
         /// <param name="functionContext">The function context.</param>
-        /// <returns>A new instance of <see cref="ConverterContext"/></returns>
-        ConverterContext Create(Type targetType, object? source, FunctionContext functionContext);
-
-        /// <summary>
-        /// Creates a new instance of <see cref="ConverterContext"/>
-        /// </summary>
-        /// <param name="targetType">The target type.</param>
-        /// <param name="source">The source.</param>
-        /// <param name="functionContext">The function context.</param>
-        /// <param name="properties">A property bag for specifying additional meta data.</param>
+        /// <param name="properties">A property bag for specifying additional metadata.</param>
         /// <returns>A new instance of <see cref="ConverterContext"/></returns>
         ConverterContext Create(Type targetType, object? source, FunctionContext functionContext, IReadOnlyDictionary<string, object> properties);
     }