+++ test/E2ETests/E2EApps/E2EApp/Blob/BlobTriggerBindingFunctions.cs
@@ -55,7 +55,7 @@ namespace Microsoft.Azure.Functions.Worker.E2EApp.Blob
             [BlobTrigger("test-trigger-stream-dotnet-isolated/{name}")] Stream stream, string name,
             FunctionContext context)
         {
-            using var blobStreamReader = new StreamReader(stream);
+            var blobStreamReader = new StreamReader(stream);
             string content = await blobStreamReader.ReadToEndAsync();
             _logger.LogInformation("StreamTriggerOutput: {c}", content);
         }