+++ sdk/release_notes.md
@@ -4,14 +4,18 @@
 - My change description (#PR/#issue)
 -->
 
-### Microsoft.Azure.Functions.Worker.Sdk <version>
+### Microsoft.Azure.Functions.Worker.Sdk <version> (meta package)
 
 - <entry>
 
-### Microsoft.Azure.Functions.Worker.Sdk.Analyzers <version>
+### Microsoft.Azure.Functions.Worker.Sdk.Analyzers <version> (delete if not updated)
 
 - <entry>
 
 ### Microsoft.Azure.Functions.Worker.Sdk.Generators <version>
 
-- <entry>
+- Parse named arguments by type (#1877)
+- Refactor source gen to walk dependent assemblies (#1896)
+- Add diagnostic descriptor logs for parsing binding arguments in source gen (#1882)
+- Use project root namespace for generated types (#1158)
+